TestDataUtils enhancements
