+++ src/org/labkey/test/util/TestDataGenerator.java
@@ -40,6 +40,7 @@ import org.labkey.serverapi.reader.TabLoader;
 import org.labkey.test.WebTestHelper;
 import org.labkey.test.params.FieldDefinition;
 import org.labkey.test.util.data.ColumnNameMapper;
+import org.labkey.test.util.data.RecordIterator;
 import org.labkey.test.util.data.TestDataUtils;
 import org.labkey.test.util.query.QueryApiHelper;
 
@@ -55,7 +56,6 @@ import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
-import java.util.NoSuchElementException;
 import java.util.Objects;
 import java.util.Set;
 import java.util.concurrent.ThreadLocalRandom;
@@ -770,7 +770,7 @@ public class TestDataGenerator
      */
     public File writeData(String fileName)
     {
-        return writeData(fileName, new FileRowIterator(getFieldsForFile(), _rows));
+        return writeData(fileName, new RecordIterator(getFieldsForFile(), _rows));
     }
 
     /**
@@ -781,7 +781,7 @@ public class TestDataGenerator
      */
     public File writeData(String fileName, int numRows)
     {
-        return writeData(fileName, new FileRowIterator(getFieldsForFile(), this::generateRow, numRows));
+        return writeData(fileName, new RecordIterator(getFieldsForFile(), this::generateRow, numRows));
     }
 
     /**
@@ -980,72 +980,4 @@ public class TestDataGenerator
         return DomainUtils.doesDomainExist(containerPath, schema, queryName);
     }
 
-}
-
-class FileRowIterator implements Iterator<List<Object>>
-{
-    private final List<String> headers;
-    private final Iterator<Map<String, Object>> rows;
-
-    private boolean firstRow = true;
-
-    public FileRowIterator(@NotNull List<String> headers, @NotNull Iterator<Map<String, Object>> rows)
-    {
-        this.headers = Objects.requireNonNull(headers);
-        this.rows = Objects.requireNonNull(rows);
-    }
-
-    public FileRowIterator(@NotNull List<String> headers, @NotNull Supplier<Map<String, Object>> rowSupplier, final int rowCount)
-    {
-        this(headers, new Iterator<>()
-        {
-            int count = 0;
-
-            @Override
-            public boolean hasNext()
-            {
-                return count < rowCount;
-            }
-
-            @Override
-            public Map<String, Object> next()
-            {
-                count++;
-                return rowSupplier.get();
-            }
-        });
-    }
-
-    public FileRowIterator(@NotNull List<String> headers, @NotNull List<Map<String, Object>> rows)
-    {
-        this(headers, rows.iterator());
-    }
-
-    @Override
-    public boolean hasNext()
-    {
-        return firstRow || rows.hasNext();
-    }
-
-    @Override
-    public List<Object> next()
-    {
-        if (!hasNext())
-            throw new NoSuchElementException();
-
-        if (firstRow)
-        {
-            firstRow = false;
-            return Collections.unmodifiableList(headers);
-        }
-        else
-        {
-            return rowMapToList(rows.next());
-        }
-    }
-
-    private List<Object> rowMapToList(Map<String, Object> row)
-    {
-        return headers.stream().map(h -> row.getOrDefault(h, "")).toList();
-    }
 }
